+++ src/main/java/com/ookawara/book/application/controller/BookController.java
@@ -1,9 +1,9 @@
-package com.ookawara.book_.application.controller;
+package com.ookawara.book.application.controller;
 
-import com.ookawara.book_.application.entity.Book;
-import com.ookawara.book_.application.entity.BookAllData;
-import com.ookawara.book_.application.entity.Category;
-import com.ookawara.book_.application.service.BookService;
+import com.ookawara.book.application.entity.BookAllData;
+import com.ookawara.book.application.entity.Category;
+import com.ookawara.book.application.entity.Book;
+import com.ookawara.book.application.service.BookService;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
