+++ src/main/java/com/ookawara/book/application/exception/BookNotFoundException.java
@@ -1,4 +1,4 @@
-package com.ookawara.book_.application.exception;
+package com.ookawara.book.application.exception;
 
 public class BookNotFoundException extends RuntimeException {
     public BookNotFoundException(String message) {
