+++ src/main/java/com/ookawara/book/application/exception/BooksOkException.java
@@ -1,4 +1,4 @@
-package com.ookawara.book_.application.exception;
+package com.ookawara.book.application.exception;
 
 public class BooksOkException extends RuntimeException {
     public BooksOkException(String message) {
