+++ src/main/java/com/ookawara/book/application/exception/CategoryNotFoundException.java
@@ -1,4 +1,4 @@
-package com.ookawara.book_.application.exception;
+package com.ookawara.book.application.exception;
 
 public class CategoryNotFoundException extends RuntimeException {
     public CategoryNotFoundException(String message) {
