+++ src/test/java/com/ookawara/book/application/BookAllDataApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ookawara.book_.application;
+package com.ookawara.book.application;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
